+++ src/classes/syncer/WebResourceSyncer.ts
@@ -18,7 +18,6 @@ export default class WebResourceSyncer {
 	}
 
 	//TODO - this should be not a singular function, but a list of steps
-	//TODO - this should not use any[]
 	private async reportProgress<T>(actionName: string, asyncFunc: Function, ...functionParams: any[]): Promise<T> {
 
 		return await vscode.window.withProgress({
@@ -135,5 +134,31 @@ export default class WebResourceSyncer {
 		};
 
 		await this.reportProgress("Publishing WebResource...", asyncFunc, path);
+	}
+
+	async testConnection(): Promise<boolean> {
+		let asyncFunc = async () => {
+			const args = ['authenticate', '--conn-string', this.connString];
+
+			const procResult = await this._execFile(this._exePath, args, {
+				shell: true,
+				windowsHide: true,
+			});
+
+			let response: WebResoureceSyncerResponse = JSON.parse(procResult.stdout);
+
+			let string = `Action: ${response.action.actionName}. Successful: ${response.action.successful}. ` 
+							+ (response.action.successful ? "" : `Error message: ${response.action.errorMessage}`);
+
+			if (response.action.successful) {
+				vscode.window.showInformationMessage(string);
+			} else {
+				vscode.window.showErrorMessage(string, "Rerun this action?");
+			}
+
+			return response.action.successful;
+		};
+
+		return await this.reportProgress<boolean>("Testing connection to Power Platform...", asyncFunc);
 	}
 }