+++ src/classes/syncer/WebResourceSyncerConfigurationManager.ts
@@ -3,13 +3,15 @@ import { WebResourceSyncerConfiguration } from './WebResourceSyncerConfiguration
 
 export abstract class WebResourceSyncerConfigurationManager {
 	public static workspaceConfigFileName: string = 'bamboo.conf.json';
+	private static undefinedWorkspaceExceptionMessage: string = "Cannot activate Bamboo. Workspace is undefined";
 
 	public static async currentWorkspaceHasConfigFile(): Promise<boolean> {
-		if (vscode.workspace.workspaceFolders === undefined) {
-			throw new Error("Cannot activate extension. Workspace is undefined");
+		let currentWorkspaceFolders = vscode.workspace.workspaceFolders;
+		if (currentWorkspaceFolders === undefined || currentWorkspaceFolders?.length > 1) {
+			throw new Error(this.undefinedWorkspaceExceptionMessage);
 		}
 
-		const workspacePath = vscode.workspace.workspaceFolders[0].uri.path;
+		const workspacePath = currentWorkspaceFolders![0].uri.path;
 
 		try {
 			await vscode.workspace.fs.stat(vscode.Uri.file(workspacePath + '/' + this.workspaceConfigFileName));
@@ -20,10 +22,12 @@ export abstract class WebResourceSyncerConfigurationManager {
 	}
 
 	public static async getConfigFileAsJson(): Promise<WebResourceSyncerConfiguration> {
-		if (vscode.workspace.workspaceFolders === undefined) {
-			throw new Error("Cannot activate extension. Workspace is undefined");
+		let currentWorkspaceFolders = vscode.workspace.workspaceFolders;
+		if (currentWorkspaceFolders === undefined || currentWorkspaceFolders?.length > 1) {
+			throw new Error(this.undefinedWorkspaceExceptionMessage);
 		}
-		const workspacePath = vscode.workspace.workspaceFolders[0].uri.path;
+
+		const workspacePath = currentWorkspaceFolders![0].uri.path;
 		const packageJsonUri = vscode.Uri.file(workspacePath + '/' + this.workspaceConfigFileName);
 		try {
 			const dataAsU8Array = await vscode.workspace.fs.readFile(packageJsonUri);
@@ -35,13 +39,27 @@ export abstract class WebResourceSyncerConfigurationManager {
 		}
 	}
 
-	public static async getWebResourceFileMapping(webResourceFilePath: string): Promise<string | null> {
+	public static async getWRPathInPowerApps(pathToFileOnDisk: string): Promise<string | null> {
 		const json: WebResourceSyncerConfiguration = await this.getConfigFileAsJson();
-		if (json.fileMappings.hasOwnProperty(webResourceFilePath)) {
-			return json.fileMappings[webResourceFilePath];
+		if (json.fileMappings.hasOwnProperty(pathToFileOnDisk)) {
+			return json.fileMappings[pathToFileOnDisk];
 		} else {
 			return null;
 		}
+	}
+
+	public static async getWRDiskPath(pathToFileInPowerApps: string): Promise<string | null> {
+		const json: WebResourceSyncerConfiguration = await this.getConfigFileAsJson();
+
+		for (let key in json.fileMappings) {
+			if (json.fileMappings.hasOwnProperty(key)) {
+				if (json.fileMappings[key] === pathToFileInPowerApps) {
+					return key;
+				}
+			}
+		}
+
+		return null;
 	}
 
 	public static async getConnectionString(): Promise<string> {
@@ -88,10 +106,11 @@ export abstract class WebResourceSyncerConfigurationManager {
 	}
 
 	private static async overwriteConfigFile(configFile: any): Promise<void> {
-		if (vscode.workspace.workspaceFolders === undefined) {
-			throw new Error("Cannot activate extension. Workspace is undefined");
+		let currentWorkspaceFolders = vscode.workspace.workspaceFolders;
+		if (currentWorkspaceFolders === undefined || currentWorkspaceFolders?.length > 1) {
+			throw new Error(this.undefinedWorkspaceExceptionMessage);
 		}
-		const workspacePath = vscode.workspace.workspaceFolders[0].uri.path;
+		const workspacePath = currentWorkspaceFolders![0].uri.path;
 		const packageJsonUri = vscode.Uri.file(workspacePath + '/' + this.workspaceConfigFileName);
 		try {
 			let configFileAsString = JSON.stringify(configFile, null, 4);