+++ CHANGELOG.md
@@ -21,3 +21,10 @@
 ## [0.2.1]
 
 - Fixed error messages
+
+## [0.2.2]
+
+- [README UPDATES](https://github.com/Root16/bamboo/issues/43)
+- [Webresource tree view lists the matching files on disk underneath the file name in PA](https://github.com/Root16/bamboo/issues/42)
+- [Authenticate automatically on startup](https://github.com/Root16/bamboo/issues/41)
+- [Fix extension failing when workspace is undefined](https://github.com/Root16/bamboo/issues/40)